package/libnetconf2: add patch to fix incompatiblity with libssh
The libnetconf2 upstream project rejects building with ssh support for
libssh 0.9.3 and 0.9.4 due to an existing regression.

This libssh regression bug is already fixed upstream and the buildroot
package by 39099153d9404cd6f107f0fd92ce8686bbd1beb2.

Thus we can drop the libnetconf2 commit that adds this libssh version to
the blacklist.

Signed-off-by: Heiko Thiery <heiko.thiery@gmail.com>